==> 01_preprocessing/04_Create_allgo_and_all_nogo_file.ipynb <==
import pandas as pd
import numpy as np
import os
import glob

path = '/media/data/HCPAging/data/3T_tfMRI_CARIT/data/'

subjects = sorted(os.listdir(path))
#print(subjects)

for subject in subjects:
    file_go = path+subject+'/EVs/go.txt'
    if os.path.getsize(file_go) != 0:
        df1 = pd.read_csv(file_go, sep=' ', header=None)
        file1 = path+subject+'/EVs/miss.txt'
        if os.path.getsize(file1) == 0:
            df1.to_csv(path+subject+'/EVs/allgo.txt', sep=' ', header=False, index=False)
            print(subject, 'There no incorrect go trials.')
        if os.path.getsize(file1) != 0:
            df2 = pd.read_csv(file1, sep=' ', header=None)
            df_go = pd.concat([df1,df2], axis=0, ignore_index=True)
            df_go = df_go.sort_values(by=0)
            df_go.to_csv(path+subject+'/EVs/allgo.txt', sep=' ', header=False, index=False)
    if os.path.getsize(file_go) == 0:
        file1 = path+subject+'/EVs/miss.txt'
        if os.path.getsize(file1) == 0:
            print(subject, "DOESN'T HAVE VALID GO FILES!!!") 
        if os.path.getsize(file1) != 0:
            print(subject, 'missed all GO trials!!!!')
            df2 = pd.read_csv(file1, sep=' ', header=None)
            df2.to_csv(path+subject+'/EVs/allgo.txt', sep=' ', header=False, index=False)

    file_nogo = path+subject+'/EVs/nogoCR.txt'
    if os.path.getsize(file_nogo) != 0:
        df3 = pd.read_csv(file_nogo, sep=' ', header=None)
        file2 = path+subject+'/EVs/nogoFA.txt'
        if os.path.getsize(file2) == 0:
            df3.to_csv(path+subject+'/EVs/allnogo.txt', sep=' ', header=False, index=False)
            print(subject, 'There no incorrect nogo trials.')
        if os.path.getsize(file2) != 0:
            df4 = pd.read_csv(file2, sep=' ', header=None)
            df_nogo = pd.concat([df3,df4], axis=0, ignore_index=True)
            df_nogo = df_nogo.sort_values(by=0)
            df_nogo.to_csv(path+subject+'/EVs/allnogo.txt', sep=' ', header=False, index=False)
    if os.path.getsize(file_nogo) == 0:
        file2 = path+subject+'/EVs/nogoFA.txt'
        if os.path.getsize(file2) == 0:
            print(subject, "DOESN'T HAVE VALID NOGO FILES!!!") 
        if os.path.getsize(file2) != 0:
            df4 = pd.read_csv(file2, sep=' ', header=None)
            df4.to_csv(path+subject+'/EVs/allnogo.txt', sep=' ', header=False, index=False)

    if os.path.getsize(file_go) == 0 and os.path.getsize(file2) == 0:
        print(subject, 'SEEMS LIKE DID NO TASK!')
