==> 01_preprocessing/04_create_distractor_3rd_ev_for_facename_task.ipynb <==
import pandas as pd
import numpy as np
import glob
import os

path = '/media/data/HCPAging/data/3T_tfMRI_FACENAME/data/'

subjects = sorted([i.split('/')[-1] for i in glob.glob(path+'*_V1_MR')])

for subject in subjects:
    #read encoding txt
    enc = pd.read_csv(path+subject+'/EVs/encoding.txt', sep=' ', header=None)
    #calculate timing for distractor
    dct_d = {}
    for ind in enc.index:
        dct_d[ind] = [(enc.loc[ind,0].round() + enc.loc[ind,1].round() + 2), 18, 1]
    #make dataframe
    dist = pd.DataFrame(dct_d).T.astype(int)
    #save to file with tab spacing
    dist.to_csv(path+subject+'/EVs/distractor.txt', index=False, header=False, sep='\t')
