==> 01_preprocessing/09_Cleaning_tables_before_ML.ipynb <==
import pandas as pd
import numpy as np
import os
import glob

path_beh = '/media/data/HCPAging/HCPAgingBehavior/'
path_data = '/media/data/HCPAging/data/'

df_cog = pd.read_csv(path_beh+'cogcomp01.txt', sep='\t')
df_cog.drop([0], inplace = True)
df_cog = df_cog[['src_subject_id', 'nih_totalcogcomp_unadjusted', 
                 'nih_crycogcomp_unadjusted','nih_fluidcogcomp_unadjusted']]
#df_cog.insert(loc=2, column='age_Yrs', value=(np.array(df_cog['interview_age'], dtype=int)/12).round())
df_cog = df_cog.rename(columns={'src_subject_id':'subject'})
df_cog.index=df_cog['subject']
df_cog = df_cog.drop('subject', axis=1)
df_cog = df_cog.sort_index(axis=0)
df_cog['nih_fluidcogcomp_unadjusted'] = df_cog['nih_fluidcogcomp_unadjusted'].replace('999', np.nan) 
df_cog = df_cog.dropna()
display(df_cog)

725-614

df_dem = pd.read_csv(path_beh+'ndar_subject01.txt', sep='\t')
print(df_dem.columns)

df_dem = pd.read_csv(path_beh+'ndar_subject01.txt', sep='\t')
df_dem.drop([0], inplace = True)
df_dem = df_dem[['src_subject_id','interview_age', 'sex', 'race','ethnic_group', 'family_user_def_id']]
df_dem = df_dem.rename(columns={'src_subject_id':'subject'})
df_dem.index=df_dem['subject']
df_dem = df_dem.drop('subject', axis=1)
df_dem = df_dem.sort_index(axis=0)
#display(df_dem)

dct_mov={}
for file in sorted(glob.glob('/media/data/HCPAging/data/3T_*/movement_*')):
    dct_mov[file.split('/')[-2].split('_')[-1]] = pd.read_csv(file, index_col=0, header=None,squeeze=True)
    dct_mov[file.split('/')[-2].split('_')[-1]].name = file.split('/')[-2].split('_')[-1]

ix = sorted(set(dct_mov['REST'].index))
for key in dct_mov.keys():
    if key != 'REST':
        ix = sorted(set(np.concatenate((ix, sorted(dct_mov[key].index)), axis=None)))  
        
df_mov = pd.DataFrame({'ix':ix}, index=ix)
for key in dct_mov.keys():
    df_mov = pd.concat([df_mov, dct_mov[key]], axis=1)
df_mov.drop(['ix'], axis=1, inplace = True)
df_mov = df_mov.dropna(axis=0)
df_mov.columns = ['rest', 'carit', 'face', 'vism']
display(df_mov)

iddx_mov = [i[:10] for i in df_mov.index]

df_mov.index = iddx_mov

df_rest = pd.read_csv('/media/data/HCPAging/data/3T_REST/FC_rest_group_z_hpass.csv', index_col=0)
df_carit1 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_CARIT/CARIT_group_table_con1.csv', index_col=0)
df_carit3 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_CARIT/CARIT_group_table_con3.csv', index_col=0)
df_carit4 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_CARIT/CARIT_group_table_con4.csv', index_col=0)
df_caritFC = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_CARIT/CARIT_task_FC.csv', index_col=0)

df_face1 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_FACENAME/FACENAME_group_table_3EV_con1.csv', index_col=0)
df_face2 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_FACENAME/FACENAME_group_table_3EV_con2.csv', index_col=0)
df_face3 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_FACENAME/FACENAME_group_table_3EV_con3.csv', index_col=0)
df_face4 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_FACENAME/FACENAME_group_table_3EV_con4.csv', index_col=0)
df_face5 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_FACENAME/FACENAME_group_table_3EV_con5.csv', index_col=0)
df_face6 = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_FACENAME/FACENAME_group_table_3EV_con6.csv', index_col=0)
df_faceFC = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_FACENAME/FACENAME_task_FC_3EV.csv', index_col=0)

df_vism = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_VISMOTOR/VISMOTOR_group_table.csv', index_col=0)
df_vismFC = pd.read_csv('/media/data/HCPAging/data/3T_tfMRI_VISMOTOR/VISMOTOR_task_FC.csv', index_col=0)


for dataset in [df_rest, df_carit1, df_carit3, df_carit4, df_caritFC, 
                df_face1, df_face2, df_face3, df_face4, df_face5, df_face6, 
                df_faceFC, df_vism, df_vismFC ]:
    dataset.index = [i[:10] for i in dataset.index]

dct_struct = {}
for file in sorted(glob.glob('/media/data/HCPAging/data/3T_Structural/*.csv')):
    data = pd.read_csv(file, index_col=0)
    iddx_d = [i[:10] for i in data.index]
    data.index = iddx_d
    dct_struct[file.split('/')[-1].split('.')[0]] = data

dct_struct.keys()

len(df_faceFC.index)

subjects = sorted(list(
set(df_mov.index) &
set(df_rest.index) &
set(df_carit1.index) &
set(df_carit3.index) &
set(df_carit4.index) &
set(df_caritFC.index) &
set(df_face1.index) &
set(df_face2.index) &
set(df_face3.index) &  
set(df_face4.index) &
set(df_face5.index) &
set(df_face6.index) &  
set(df_faceFC.index) &     
set(df_vism.index) &
set(df_vismFC.index) &   
set(dct_struct['cortical_area'].index) &
set(dct_struct['cortical_thickness'].index) &
set(dct_struct['subcortical_volume'].index) &
set(dct_struct['total_brain_volume'].index) 
))
print(len(subjects))

'''subjects = sorted(list(
set(df_mov.index) &
set(df_rest.index) &
set(df_carit.index) &
set(df_face.index) &
set(df_vism.index) &
set(dct_struct['cortical_area'].index) &
set(dct_struct['cortical_thickness'].index) &
set(dct_struct['subcortical_volume'].index) &
set(dct_struct['total_brain_volume'].index) 
))
print(len(subjects))'''
#710

QS_tab = pd.read_csv('/media/data/HCPAging/Doc/HCA_LS_2.0_subject_completeness.csv')[['src_subject_id', 'QC_Issue_Codes']]
QS_tab.drop([0], axis=0, inplace = True)
display(QS_tab)
subj_to_del = QS_tab[QS_tab['QC_Issue_Codes'].isin(['A', 'B1a', 'B1b', 'B1c', 'B2a', 'B2b', 'B2c', 'B2d', 'B3'])]['src_subject_id']

QS_tab[QS_tab['QC_Issue_Codes'].isin(['A', 'B1a', 'B1b', 'B1c', 'B2a', 'B2b', 'B2c', 'B2d', 'B3'])].dropna()

for sd in subj_to_del:
    #sd = sd+'_V1_MR'
    if sd in subjects:
        #print(sd,' in subj')
        subjects.remove(sd)

print(len(subjects))

710-595


cog_ind =df_cog.dropna().index#[ i+'_V1_MR' for i in df_cog.dropna().index]

subjects = sorted(set(cog_ind) & set(subjects))
len(subjects)

path_out = '/media/data/HCPAging/data/MLTablesMultCon/'

datasets = [df_mov,
df_rest,
df_carit1,
df_carit3,
df_carit4,
df_caritFC,
df_face1,
df_face2,
df_face3,
df_face4,
df_face5,
df_face6, 
df_faceFC,    
df_vism,
df_vismFC,
dct_struct['cortical_area'],
dct_struct['cortical_thickness'],
dct_struct['subcortical_volume'],
dct_struct['total_brain_volume'],
df_dem,
df_cog]

name_dsets = ['movements.csv',
'rest_hpass.csv',
'carit-con1.csv',
'carit-con3.csv',
'carit-con4.csv',
'CARIT_taskFC.csv',
'FACENAME_group_table_3EV_con1.csv',
'FACENAME_group_table_3EV_con2.csv',
'FACENAME_group_table_3EV_con3.csv',
'FACENAME_group_table_3EV_con4.csv',
'FACENAME_group_table_3EV_con5.csv',
'FACENAME_group_table_3EV_con6.csv',
'FACENAME_task_FC_3EV.csv',
'vism.csv',
'VISMOTOR_taskFC.csv',
'surf.csv',
'cort.csv',
'subc.csv',
'VolBrain.csv',
'demography.csv',
'cognition.csv']

for dataset, name_dset in zip(datasets, name_dsets):
    dataset.reindex(index=subjects).to_csv(path_out+name_dset)


ffdem = df_dem.reindex(index=subjects)
#display(ffdem)

ffdem['interview_age'].astype('int').mean()

ffdem['interview_age'].astype('int').mean() / 12

ffdem['interview_age'].astype('int').std()

ffdem['interview_age'].astype('int').std() / 12

len(ffdem['sex'][ffdem['sex'] == 'F'])
